refactor: remove "port" package from core
