+++ dataplane-sdk-core/src/main/java/org/eclipse/dataplane/Dataplane.java
@@ -34,6 +34,12 @@ import org.eclipse.dataplane.domain.registration.Authorization;
 import org.eclipse.dataplane.domain.registration.AuthorizationProfile;
 import org.eclipse.dataplane.domain.registration.ControlPlaneRegistrationMessage;
 import org.eclipse.dataplane.domain.registration.DataPlaneRegistrationMessage;
+import org.eclipse.dataplane.exception.AuthorizationNotSupported;
+import org.eclipse.dataplane.exception.ControlPlaneNotRegistered;
+import org.eclipse.dataplane.exception.DataFlowNotifyControlPlaneFailed;
+import org.eclipse.dataplane.exception.DataplaneNotRegistered;
+import org.eclipse.dataplane.exception.ResourceNotFoundException;
+import org.eclipse.dataplane.exception.UnauthorizedException;
 import org.eclipse.dataplane.logic.OnCompleted;
 import org.eclipse.dataplane.logic.OnPrepare;
 import org.eclipse.dataplane.logic.OnResume;
@@ -41,17 +47,11 @@ import org.eclipse.dataplane.logic.OnStart;
 import org.eclipse.dataplane.logic.OnStarted;
 import org.eclipse.dataplane.logic.OnSuspend;
 import org.eclipse.dataplane.logic.OnTerminate;
-import org.eclipse.dataplane.port.exception.AuthorizationNotSupported;
-import org.eclipse.dataplane.port.exception.ControlPlaneNotRegistered;
-import org.eclipse.dataplane.port.exception.DataFlowNotifyControlPlaneFailed;
-import org.eclipse.dataplane.port.exception.DataplaneNotRegistered;
-import org.eclipse.dataplane.port.exception.ResourceNotFoundException;
-import org.eclipse.dataplane.port.exception.UnauthorizedException;
-import org.eclipse.dataplane.port.store.ControlPlaneStore;
-import org.eclipse.dataplane.port.store.DataFlowStore;
-import org.eclipse.dataplane.port.store.InMemoryControlPlaneStore;
-import org.eclipse.dataplane.port.store.InMemoryDataFlowStore;
-import org.eclipse.dataplane.port.store.Stores;
+import org.eclipse.dataplane.store.ControlPlaneStore;
+import org.eclipse.dataplane.store.DataFlowStore;
+import org.eclipse.dataplane.store.InMemoryControlPlaneStore;
+import org.eclipse.dataplane.store.InMemoryDataFlowStore;
+import org.eclipse.dataplane.store.Stores;
 
 import java.net.URI;
 import java.net.http.HttpClient;
