+++ dataplane-sdk-core/src/main/java/org/eclipse/dataplane/exception/ControlPlaneNotRegistered.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2025 Think-it GmbH
+ *  Copyright (c) 2026 Think-it GmbH
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -12,10 +12,12 @@
  *
  */
 
-package org.eclipse.dataplane.port.exception;
+package org.eclipse.dataplane.exception;
 
-public class DataplaneNotRegistered extends RuntimeException {
-    public DataplaneNotRegistered(String message) {
-        super(message);
+public class ControlPlaneNotRegistered extends Exception {
+
+    public ControlPlaneNotRegistered(String controlplaneId) {
+        super("Controlplane " + controlplaneId + " not registered");
     }
+
 }
