+++ dataplane-sdk-core/src/main/java/org/eclipse/dataplane/exception/IllegalAttributeTypeException.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataplane.port.exception;
+package org.eclipse.dataplane.exception;
 
 public class IllegalAttributeTypeException extends RuntimeException {
     public IllegalAttributeTypeException(String message) {
