+++ dataplane-sdk-core/src/main/java/org/eclipse/dataplane/exception/PersistenceException.java
@@ -12,12 +12,19 @@
  *
  */
 
-package org.eclipse.dataplane.port.exception;
+package org.eclipse.dataplane.exception;
 
-public class UnauthorizedException extends RuntimeException {
+/**
+ * Indicates an error during database interactions, i.e. an error occurred persisting, reading or
+ * deleting an entry.
+ */
+public class PersistenceException extends RuntimeException {
 
-    public UnauthorizedException(String message) {
+    public PersistenceException(String message) {
         super(message);
     }
 
+    public PersistenceException(String message, Throwable cause) {
+        super(message, cause);
+    }
 }
