+++ dataplane-sdk-core/src/test/java/org/eclipse/dataplane/store/InMemoryControlPlaneStoreTest.java
@@ -15,8 +15,6 @@
 package org.eclipse.dataplane.store;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.eclipse.dataplane.port.store.ControlPlaneStore;
-import org.eclipse.dataplane.port.store.InMemoryControlPlaneStore;
 
 import static com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES;
