+++ dataplane-sdk-postgresql/src/main/java/org/eclipse/dataplane/store/postgresql/AbstractSqlStore.java
@@ -18,7 +18,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.eclipse.dataplane.port.exception.PersistenceException;
+import org.eclipse.dataplane.exception.PersistenceException;
 
 import java.sql.Connection;
 import javax.sql.DataSource;
