+++ dataplane-sdk-postgresql/src/main/java/org/eclipse/dataplane/store/postgresql/PostgresControlPlaneStore.java
@@ -18,9 +18,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.eclipse.dataplane.domain.Result;
 import org.eclipse.dataplane.domain.controlplane.ControlPlane;
 import org.eclipse.dataplane.domain.registration.AuthorizationProfile;
-import org.eclipse.dataplane.port.exception.PersistenceException;
-import org.eclipse.dataplane.port.exception.ResourceNotFoundException;
-import org.eclipse.dataplane.port.store.ControlPlaneStore;
+import org.eclipse.dataplane.exception.PersistenceException;
+import org.eclipse.dataplane.exception.ResourceNotFoundException;
+import org.eclipse.dataplane.store.ControlPlaneStore;
 
 import java.net.URI;
 import javax.sql.DataSource;
