+++ dataplane-sdk-postgresql/src/test/java/org/eclipse/dataplane/store/postgresql/PostgresDataFlowStoreTest.java
@@ -15,7 +15,7 @@
 package org.eclipse.dataplane.store.postgresql;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.eclipse.dataplane.port.store.DataFlowStore;
+import org.eclipse.dataplane.store.DataFlowStore;
 import org.eclipse.dataplane.store.DataFlowStoreTestBase;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
