+++ e2e-tests/src/test/java/org/eclipse/dataplane/DataplaneTest.java
@@ -20,9 +20,9 @@ import org.eclipse.dataplane.domain.Result;
 import org.eclipse.dataplane.domain.dataflow.DataFlowPrepareMessage;
 import org.eclipse.dataplane.domain.registration.AuthorizationProfile;
 import org.eclipse.dataplane.domain.registration.ControlPlaneRegistrationMessage;
-import org.eclipse.dataplane.port.exception.DataFlowNotifyControlPlaneFailed;
-import org.eclipse.dataplane.port.exception.DataplaneNotRegistered;
-import org.eclipse.dataplane.port.exception.ResourceNotFoundException;
+import org.eclipse.dataplane.exception.DataFlowNotifyControlPlaneFailed;
+import org.eclipse.dataplane.exception.DataplaneNotRegistered;
+import org.eclipse.dataplane.exception.ResourceNotFoundException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
