+++ e2e-tests/src/test/java/org/eclipse/dataplane/api/ControlPlaneRegistrationApiTest.java
@@ -20,8 +20,8 @@ import org.eclipse.dataplane.HttpServer;
 import org.eclipse.dataplane.authorization.TestAuthorization;
 import org.eclipse.dataplane.domain.registration.AuthorizationProfile;
 import org.eclipse.dataplane.domain.registration.ControlPlaneRegistrationMessage;
+import org.eclipse.dataplane.exception.ResourceNotFoundException;
 import org.eclipse.dataplane.port.DataPlaneRegistrationApiController;
-import org.eclipse.dataplane.port.exception.ResourceNotFoundException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
