+++ e2e-tests/src/test/java/org/eclipse/dataplane/scenario/AuthorizationTest.java
@@ -21,8 +21,8 @@ import org.eclipse.dataplane.authorization.TestAuthorization;
 import org.eclipse.dataplane.domain.Result;
 import org.eclipse.dataplane.domain.dataflow.DataFlowStatusMessage;
 import org.eclipse.dataplane.domain.registration.ControlPlaneRegistrationMessage;
+import org.eclipse.dataplane.exception.DataFlowNotifyControlPlaneFailed;
 import org.eclipse.dataplane.port.DataPlaneSignalingApiController;
-import org.eclipse.dataplane.port.exception.DataFlowNotifyControlPlaneFailed;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
